Fix JSON decoder bug when events are coming in on STDOUT
When events come in on stdout they are formatted with:

    T2-HARNESS-EVENT: {...json blob...}

This is done in _process_stdout_line:

    my $event_data = decode_json($json);
    # Do some manipulations with $event_data
    push @event_datas => $event_data;

@event_datas eventually ends up getting placed on the event buffer.

Cool.

However -- later on we poll the event buffer in _poll_event and expect
to be able to decode_json each item in the buffer.  Whoops!  Some of
these might already be hashrefs.

Conditionally check the type -- if its a hashref, we have already done
the json decode, otherwise run through the json decode.