Pass T2_STREAM_FILE along with T2_FORMATTER in IPC
When we run child processes with streams enabled, we pass a few
settings:

1.) -MTest2::Formatter::Stream=file,<event_file>
2.) $ENV{T2_FORMATTER} = "Stream"

This works out pretty well when we are running a straight perl
program.  However, if the perl program were to, for example, re-exec
itself and lose the Test2::Formatter::Stream import, all it ends up
with is T2_FORMATTER.  At this point, we have no choice but to write
stream events to stdout - SAD.

Luckily Test2::Formatter::Stream already supports another environment
variable for controlling the output file - T2_STREAM_FILE.  Let's pass
that too.