Fix empty lines when using STDOUT Stream output
Many cases when we output Stream events to stdout, we'll end up
outputting something like this:

    ^Regular line of output$
    ^T2-HARNESS-EVENT: 1 {... json blob ...}$

While processing, we'd consume the line and end up with something like
this:

    ^Regular line of output$
    ^$

There is zero reason to keep around this empty line - just drop it
from the processing.  If we don't we get a bunch of:

    ( STDOUT )
    ( STDOUT )
    ( STDOUT )

lines while running a test where events have been stripped.

In other cases we might have something like this:

    ^Regular line of output$
    ^SomethingT2-HARNESS-EVENT: 1 {... json blob ...}$

In this case we'd process and end up with:

    ^Regular line of output$
    ^Something$

This can be passed through (although I question if this really happens
often in practice and still ends up with some minor corruption in the
form of extra newlines).