Better buffer fill logic
The buffer filling logic would not add anything to the buffers if the
combined buffer total was greater than max. This could be a problem in
cases where stderr/stdout are waiting for an event, but together meet
the max buffer fill, thus the event buffer never gets the next event
resulting in deadlock.

Now the logic fills each buffer to max instead of combined to max.