Add YATH_PERSISTENCE_DIR as an environment variable to control persistence file
The goal of this change is to be able to use a persistent test runner in an
environment where the code and tests are on a read only file system and a
writable file system (temp or otherwise) is in a different path structure

When YATH_PERSISTENCE_DIR is set the .yath-persistence.json file will be
placed within the specified path. This commit does not add assertions to
make sure this directory exists. Instead it will default to the standard
search logic that existed before this change was added.

Tests have been added and adjusted as necessary in response to this change.
These tests ghave their own specific needs (like writing to the File System)
so where necessary a local mutation of the env var was done. This was made
so that yath can test itself when the variable is set.

The variable may be useful in further tests for improving or ensuring
idempotency of the tests against the local file system by setting it to a test
harness temp directory but the author declines to make such a sweeping change
at this time.